+++ week-0010-lecture/src/main/kotlin/com/hhplus/lecture/application/ApplyHistoryManager.kt
@@ -1,11 +1,11 @@
-package com.hhplus.lecture.domain
+package com.hhplus.lecture.application
 
+import com.hhplus.lecture.application.repository.ApplyHistoryRepository
 import com.hhplus.lecture.common.exception.errors.ApplyHistoryException
 import com.hhplus.lecture.common.type.ApplyStatus
-import com.hhplus.lecture.infra.entity.ApplyHistory
-import com.hhplus.lecture.infra.entity.Lecture
-import com.hhplus.lecture.infra.entity.Student
-import com.hhplus.lecture.infra.repository.ApplyHistoryRepository
+import com.hhplus.lecture.domain.entity.ApplyHistory
+import com.hhplus.lecture.domain.entity.Lecture
+import com.hhplus.lecture.domain.entity.Student
 import org.springframework.stereotype.Component
 
 @Component
