+++ week-0010-lecture/src/main/kotlin/com/hhplus/lecture/application/LectureManager.kt
@@ -1,8 +1,8 @@
-package com.hhplus.lecture.domain
+package com.hhplus.lecture.application
 
+import com.hhplus.lecture.application.repository.LectureRepository
 import com.hhplus.lecture.common.exception.errors.LectureException
-import com.hhplus.lecture.infra.entity.Lecture
-import com.hhplus.lecture.infra.repository.LectureRepository
+import com.hhplus.lecture.domain.entity.Lecture
 import org.springframework.stereotype.Component
 import java.time.LocalDateTime
