+++ week-0010-lecture/src/main/kotlin/com/hhplus/lecture/application/StudentManager.kt
@@ -1,8 +1,8 @@
-package com.hhplus.lecture.domain
+package com.hhplus.lecture.application
 
+import com.hhplus.lecture.application.repository.StudentRepository
 import com.hhplus.lecture.common.exception.errors.StudentException
-import com.hhplus.lecture.infra.entity.Student
-import com.hhplus.lecture.infra.repository.StudentRepository
+import com.hhplus.lecture.domain.entity.Student
 import org.springframework.stereotype.Component
 
 @Component
