+++ week-0010-lecture/src/main/kotlin/com/hhplus/lecture/application/dto/LectureApplyServiceDto.kt
@@ -1,0 +1,24 @@
+package com.hhplus.lecture.application.dto
+
+object LectureApplyServiceDto {
+    data class Apply(
+        val studentId: Long,
+        val lectureId: Long,
+    )
+
+    data class ApplyResult(
+        val result: Boolean,
+        val lectureTitle: String? = null,
+    )
+
+    data class Student(
+        val studentId: Long,
+        val name: String,
+    )
+
+    data class Lecture(
+        val lectureId: Long,
+        val title: String,
+        val capacity: Int,
+    )
+}
