+++ week-0010-lecture/src/main/kotlin/com/hhplus/lecture/controller/request/LectureApplyRequest.kt
@@ -1,6 +1,6 @@
 package com.hhplus.lecture.controller.request
 
-import com.hhplus.lecture.domain.dto.LectureApplyServiceDto
+import com.hhplus.lecture.application.dto.LectureApplyServiceDto
 
 object LectureApplyRequest {
     data class Apply(
@@ -8,12 +8,11 @@ object LectureApplyRequest {
         val lectureId: Long,
     ) {
         companion object {
-            fun toApplyDto(request: Apply): LectureApplyServiceDto.Apply {
-                return LectureApplyServiceDto.Apply(
+            fun toApplyDto(request: Apply): LectureApplyServiceDto.Apply =
+                LectureApplyServiceDto.Apply(
                     studentId = request.studentId,
                     lectureId = request.lectureId,
                 )
-            }
         }
     }
 }
