+++ week-0010-lecture/src/main/kotlin/com/hhplus/lecture/controller/response/LectureApplyResponse.kt
@@ -1,6 +1,6 @@
 package com.hhplus.lecture.controller.response
 
-import com.hhplus.lecture.domain.dto.LectureApplyServiceDto
+import com.hhplus.lecture.application.dto.LectureApplyServiceDto
 
 object LectureApplyResponse {
     data class ApplyResult(
