+++ week-0010-lecture/src/main/kotlin/com/hhplus/lecture/infra/repository/implement/ApplyHistoryRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.hhplus.lecture.infra.repository.implement
 
-import com.hhplus.lecture.infra.entity.ApplyHistory
-import com.hhplus.lecture.infra.repository.ApplyHistoryRepository
+import com.hhplus.lecture.application.repository.ApplyHistoryRepository
+import com.hhplus.lecture.domain.entity.ApplyHistory
 import com.hhplus.lecture.infra.repository.jpa.JpaApplyHistoryRepository
 import org.springframework.stereotype.Repository
