+++ week-0010-lecture/src/main/kotlin/com/hhplus/lecture/infra/repository/implement/StudentRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.hhplus.lecture.infra.repository.implement
 
-import com.hhplus.lecture.infra.entity.Student
-import com.hhplus.lecture.infra.repository.StudentRepository
+import com.hhplus.lecture.application.repository.StudentRepository
+import com.hhplus.lecture.domain.entity.Student
 import com.hhplus.lecture.infra.repository.jpa.JpaStudentRepository
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Repository
