+++ week-0010-lecture/src/main/kotlin/com/hhplus/lecture/infra/repository/jpa/JpaApplyHistoryRepository.kt
@@ -1,6 +1,6 @@
 package com.hhplus.lecture.infra.repository.jpa
 
-import com.hhplus.lecture.infra.entity.ApplyHistory
+import com.hhplus.lecture.domain.entity.ApplyHistory
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface JpaApplyHistoryRepository : JpaRepository<ApplyHistory, Long> {
