+++ week-0010-lecture/src/main/kotlin/com/hhplus/lecture/infra/repository/jpa/JpaLectureRepository.kt
@@ -1,6 +1,6 @@
 package com.hhplus.lecture.infra.repository.jpa
 
-import com.hhplus.lecture.infra.entity.Lecture
+import com.hhplus.lecture.domain.entity.Lecture
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface JpaLectureRepository : JpaRepository<Lecture, Long>
