+++ week-0010-lecture/src/main/kotlin/com/hhplus/lecture/infra/repository/jpa/JpaStudentRepository.kt
@@ -1,6 +1,6 @@
 package com.hhplus.lecture.infra.repository.jpa
 
-import com.hhplus.lecture.infra.entity.Student
+import com.hhplus.lecture.domain.entity.Student
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface JpaStudentRepository : JpaRepository<Student, Long>
