+++ week-0010-lecture/src/test/kotlin/com/hhplus/lecture/integration/LectureApplyIntegrationTest.kt
@@ -1,0 +1,8 @@
+package com.hhplus.lecture.integration
+
+/**
+ * 통합테스트
+ * 1. 전체 bean 에 대한 검증
+ * 2. 선착순 동시성에 대한 검증
+ */
+class LectureApplyIntegrationTest
